+++ project_info/changelog.md
@@ -1,8 +1,27 @@
 # Change Logs
 
+## Develop Branch
 <details>
 <summary> 23.07.04 </summary>
 
 <!-- summary 아래 한칸 공백 두어야함 -->
 - 탬플릿 제작
+</details>
+
+<details>
+<summary> 23.07.05 </summary>
+
+<!-- summary 아래 한칸 공백 두어야함 -->
+- naver news crawler 완료
+- naver news crawler 세이브 기능 추가
+- mt5 요약 모델 모듈화 완료
+    - 기사 요약 테스트 확인
+</details>
+
+<details>
+<summary> 23.07.06 </summary>
+
+<!-- summary 아래 한칸 공백 두어야함 -->
+- 사람인 크롤러 모듈(셀레늄) 완성
+    - 추후 리퀘스츠로 변경요망
 </details>